Enhance postpone with UMD pattern.
This commit enhances postpone with a Universal Module Definition
pattern. Employing this pattern means that postpone can now be used
easily in none CJS-style environments and can be used as a stand-alone
module as a browser global.

This commit also removes the `docs` directory; the code is thoroughly
documented and interested users can generate docs for it using jsdoc
and similar tools.