Account for case where all elements are rerendered
This commit accounts for an edge case where Postpone would not
correctly load elemnts when they had scrolled into view. Specifically,
it takes care of the case when a set of already postponed elements are
completely re-rendered by an application. Re-rendering a page with the
same content may result in completely equivalent markup to the markup
that existed before, but from a browser's point of view, the DOM nodes
represented by those sets of markup are not identical. Postpone was
treating these nodes as identical, so if your Backbone app re-rendered a
bunch of still un-loaded images, Postpone would be unable to load them.